Reread cached properties from time to time
There was a problem where, after upgrading otter to v2, there was a chance
of properties being just cached and not reread. The reason we couldn't use
otter's magic reloading is that there's more business logic that needs to
be executed than just to reread value from DB.

One drawback is in order to keep otter's package "hidden", a clone of Loader
had to be created in common.

Additionally Verify API of server was refactored so we return a clear error
externally in case of auth failure (related to "delayed" API key check)